+++ src/main/java/roomescape/member/dto/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.dto;
 
 public record LoginMember(Long id, String name, String email, String role) {
 }
